Fixed heightmap baker

+++ gaiku-3d/src/bakers/heightmap.rs
@@ -18,10 +18,10 @@ impl Baker for HeightMapBaker {
                     continue;
                 }
                 let fz = y as f32;
-                let lb = (chunk.get(x, y, 0) * height) as f32 / 255.0;
-                let lf = (chunk.get(x, y + 1, 0) * height) as f32 / 255.0;
-                let rb = (chunk.get(x + 1, y, 0) * height) as f32 / 255.0;
-                let rf = (chunk.get(x + 1, y + 1, 0) * height) as f32 / 255.0;
+                let lb = (chunk.get(x, y, 0) as u32 * height) as f32 / 255.0;
+                let lf = (chunk.get(x, y + 1, 0) as u32 * height) as f32 / 255.0;
+                let rb = (chunk.get(x + 1, y, 0) as u32 * height) as f32 / 255.0;
+                let rf = (chunk.get(x + 1, y + 1, 0) as u32 * height) as f32 / 255.0;
 
                 let left_back = Self::index(&mut vertices_cache, [fx - 0.5, lb, fz - 0.5].into());
                 let right_back = Self::index(&mut vertices_cache, [fx + 0.5, rb, fz - 0.5].into());
